% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LinkOrgs_BuildBackend.R
\name{BuildBackend}
\alias{BuildBackend}
\title{Build the environment for LinkOrgs machine learning models. Builds a conda environment in which jax, optax, equinox, and jmp are installed.}
\usage{
BuildBackend(conda_env = "LinkOrgsEnv", conda = "auto", tryMetal = T)
}
\arguments{
\item{conda_env}{(default = \code{"LinkOrgsEnv"}) Name of the conda environment in which to place the backends.}

\item{conda}{(default = \code{auto}) The path to a conda executable. Using \code{"auto"} allows reticulate to attempt to automatically find an appropriate conda binary.}
}
\value{
Builds the computational environment for \code{LinkOrgs}. This function requires an Internet connection.
You may find out a list of conda Python paths via: \code{system("which python")}
}
\description{
Build the environment for LinkOrgs machine learning models. Builds a conda environment in which jax, optax, equinox, and jmp are installed.
}
\examples{
# For a tutorial, see
# github.com/cjerzak/linkorgs-software/

}
