% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LinkOrgsBuildML_v2_Transfer.R
\name{BuildTransfer}
\alias{BuildTransfer}
\title{A primarily internal function which builds the organizational record linkage models used in Libgober and Jerzak (2023+).}
\usage{
BuildTransfer()
}
\description{
A primarily internal function which builds the organizational record linkage models used in Libgober and Jerzak (2023+).
}
