% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LinkOrgsMain.R
\name{FastFuzzyMatch}
\alias{FastFuzzyMatch}
\title{FastFuzzyMatch}
\usage{
FastFuzzyMatch(x,y,by,...)
}
\arguments{
\item{x, y}{data frames to be merged}

\item{by, by.x, by.y}{specifications of the columns used for merging. We follow the general syntax of \code{base::merge}; see \code{?base::merge} for more details.}

\item{...}{For additional options, see ``Details''.}
}
\value{
z The merged data frame.
}
\description{
Performs fast fuzzy matching of strings based on the string distance measure specified in \code{DistanceMeasure}. Matching is parallelized using all available CPU cores to increase execution speed.
}
\details{
\code{FastFuzzyMatch} can automatically process the \code{by} text for each dataset. Users may specify the following options:
\itemize{
\item Set \code{DistanceMeasure} to control algorithm for computing pairwise string distances. Options include "\code{osa}", "\code{jaccard}", "\code{jw}". See \code{?stringdist::stringdist} for all options. (Default is "\code{jaccard}")
\item Set \code{MaxDist} to control the maximum allowed distance between two matched strings
\item Set \code{AveMatchNumberPerAlias} to control the maximum allowed distance between two matched strings. Takes priority over \code{MaxDist} if both specified.
\item Set \code{qgram} to control the character-level q-grams used in the distance measure. (Default is \code{2})
\item Set \code{RemoveCommonWords} to TRUE to remove common words (those appearing in > 10\% of aliases). (Default is \code{FALSE})
\item Set \code{NormalizeSpaces} to TRUE to remove hanging whitespaces. (Default is \code{TRUE})
\item Set \code{RemovePunctuation} to TRUE to remove punctuation. (Default is \code{TRUE})
\item Set \code{ToLower} to TRUE to ignore case. (Default is \code{TRUE})
}
}
\examples{

#Create synthetic data
x_orgnames <- c("apple","oracle","enron inc.","mcdonalds corporation")
y_orgnames <- c("apple corp","oracle inc","enron","mcdonalds co")
x <- data.frame("orgnames_x"=x_orgnames)
y <- data.frame("orgnames_y"=y_orgnames)
z <- data.frame("orgnames_x"=x_orgnames[1:2], "orgnames_y"=y_orgnames[1:2])
z_true <- data.frame("orgnames_x"=x_orgnames, "orgnames_y"=y_orgnames)

# Perform merge
linkedOrgs_fuzzy <- FastFuzzyMatch(x = x,
                       y = y,
                       by.x = "orgnames_x",
                       by.y = "orgnames_y")

}
