% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LinkOrgs_PFuzzyDiscreteFullFxn.R
\name{pFuzzyMatch_discrete}
\alias{pFuzzyMatch_discrete}
\title{pFuzzyMatch_discrete}
\usage{
pFuzzyMatch_discrete(
  x = NULL,
  by.x = NULL,
  embedx = NULL,
  y = NULL,
  by.y = NULL,
  embedy = NULL,
  embedDistMetric = NULL,
  MaxDist = NULL,
  qgram = 2,
  DistanceMeasure = "jaccard",
  AveMatchNumberPerAlias = NULL,
  nCores = NULL,
  ...
)
}
\arguments{
\item{x, y}{data frames to be merged}
}
\value{
...
}
\description{
Implements
}
\details{
...
}
\examples{

#Create synthetic data
x_orgnames <- c("apple","oracle","enron inc.","mcdonalds corporation")
y_orgnames <- c("apple corp","oracle inc","enron","mcdonalds co")
x <- data.frame("orgnames_x"=x_orgnames)
y <- data.frame("orgnames_y"=y_orgnames)

}
