% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LinkOrgs_BuildHelpers.R
\name{print2}
\alias{print2}
\title{url2dt}
\usage{
print2(text, quiet = F)
}
\arguments{
\item{text}{Text}
}
\value{
\code{...} Prints...
}
\description{
print2
}
\details{
...
}
\examples{

# Example download
print2("Hello world!")

}
