% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LinkOrgsTrainML_v1_MultiLSTM.R
\name{TrainML}
\alias{TrainML}
\title{TrainML}
\usage{
TrainML(...)
}
\description{
Internal function that performs the training of the machine learning models used for organizational record linkage algorithms of Libgober and Jerzak.
}
