% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MainFunctions.R
\name{LinkOrgs}
\alias{LinkOrgs}
\title{LinkOrgs}
\usage{
LinkOrgs(x, y, by ...)
}
\arguments{
\item{x, y}{data frames to be merged}

\item{by, by.x, by.y}{character vector(s) that specify the column names used for merging data frames \code{x} and \code{y}. The merging variables should be organizational names. See \code{?base::merge} for more details regarding syntax.}

\item{algorithm}{character; specifies which algorithm described in Jerzak and Libgober (2021) should be used. Options are "\code{markov}", "\code{bipartite}", and "\code{ml}". Default is "\code{ml}", which uses a machine learning approach to predicting the match probability.}

\item{control}{A list specifying how to process the alias text and how to compute string distances. See
``Details''.}
}
\value{
\code{z} The merged data frame.
}
\description{
Implements the organizational record linkage algorithms of Jerzak and Libgober (2021).
}
\details{
\code{LinkOrgs}automatically processes the name text for each dataset (specified by \code{by}, \code{by.x}, and/or \code{by.y}. Users may specify the following options in the \code{control} list:
\itemize{
\item Set \code{DistanceMeasure} to control algorithm for computing pairwise string distances. Options include "\code{osa}", "\code{jaccard}", "\code{jw}". See \code{?stringdist::stringdist} for all options. (Default is "\code{jaccard}")
\item Set \code{FuzzyThreshold} to control the maximum allowed distance between two matched strings
\item Set \code{qgram} to control the character-level q-grams used in the distance measure. (Default is \code{2})
\item Set \code{RemoveCommonWords} to \code{TRUE} to remove common words (those appearing in >
10\% of aliases). (Default is \code{FALSE})
\item Set \code{NormalizeSpaces} to \code{TRUE} to remove hanging whitespaces. (Default is \code{TRUE})
\item Set \code{RemovePunctuation} to \code{TRUE} to remove punctuation. (Default is \code{TRUE})
\item Set \code{ToLower} to \code{TRUE} to ignore case. (Default is \code{TRUE})
}
}
