% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LinkOrgsHelpers.R
\name{url2dt}
\alias{url2dt}
\title{url2dt}
\usage{
url2dt(url, target_extension = ".csv")
}
\arguments{
\item{url}{character string with the URL housing the data object.}

\item{target_extension}{(default = \code{".csv"}) character string describing
the target extension of the file in the downloaded .zip folder.}
}
\value{
\code{z} The downloaded data object from the URL.
}
\description{
Downloads a .zip file from a URL as a data.table from a URL.
}
\details{
\code{url2dt} downloads a zipped \code{.csv} file and loads it into memory based on the input URL.
}
\examples{

# Example download
my_dt <- url2dt(url="https://www.dropbox.com/s/iqf9ids77dckopf/Directory_LinkIt_bipartite_Embeddings.csv.zip?dl=0")

}
